Backend/requirements.txt
langgraph
langchain
langchain-community
langserve
langsmith
langchain_groq
dotenv
aiofiles
streamlit
graphviz
pillow
pdfminer.six
numpy 
pandas
elasticsearch
python-jose[cryptography]
passlib[bcrypt]
python-multipart
uvicorn[standard]
sqlalchemy
fastapi